# Conflicts:
#	src/main/java/com/example/project_client/HelloApplication.java
#	src/main/java/com/example/project_client/router/Pages.java

+++ src/main/java/com/example/project_client/view/controller/Truong/productView.java
@@ -7,13 +7,13 @@ import com.example.project_client.router.Router;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import lombok.Getter;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
 
 public final class productView {
     private static Product product;
@@ -30,7 +30,7 @@ public final class productView {
     @FXML
     private TableColumn<Product, String> image;
     @Getter
-    private List<Product> products;
+    private static List<Product> products;
     @FXML
     private ObservableList<Product> tableList;
     @FXML
@@ -40,6 +40,7 @@ public final class productView {
         try {
             setTableView();
             setColumn();
+            product = products.get(products.size() - 1);
         }
         catch (Exception e){
             System.out.println("ERROR");
@@ -50,6 +51,19 @@ public final class productView {
         Router.switchTo(Pages.MAIN_VIEW);
     }
     @FXML
+    private void readProduct(){
+        try {
+            product = tableView.getSelectionModel().getSelectedItem();
+            if(product == null) {
+                throw new Exception("Please choose Product to view information");
+            }
+            Router.switchTo(Pages.READ_PRODUCT);
+        }
+        catch (Exception e){
+            raiseAlert(e.getMessage());
+        }
+    }
+    @FXML
     private void addProduct() throws IOException {
         Router.switchTo(Pages.ADD_PRODUCT);
     }
@@ -57,29 +71,70 @@ public final class productView {
     private void changeProduct() {
         try {
             product = tableView.getSelectionModel().getSelectedItem();
+            if(product == null) {
+                throw new Exception("Please choose Product to change");
+            }
             Router.switchTo(Pages.CHANGE_PRODUCT);
         }
         catch (Exception e){
-            System.out.println("ERROR");
+            raiseAlert(e.getMessage());
         }
     }
     @FXML
     private void deleteProduct() {
         try {
             product = tableView.getSelectionModel().getSelectedItem();
-            ProductRepository.deleteProduct(product.getId().toString());
-            tableList.remove(product);
+            if(product == null){
+                throw new Exception("Please choose Product to delete");
+            }
+            askDelete();
         }
         catch (Exception e){
-            System.out.println("Please choose Product to delete");
+            raiseAlert(e.getMessage());
         }
+    }
+    private void askDelete() {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation");
+        alert.setHeaderText("Do you want to delete Product");
+        alert.setContentText("choose your option");
+
+        ButtonType buttonTypeYes = new ButtonType("Yes", ButtonBar.ButtonData.YES);
+        ButtonType buttonTypeNo = new ButtonType("No", ButtonBar.ButtonData.NO);
+
+        alert.getButtonTypes().setAll(buttonTypeYes, buttonTypeNo);
+
+        Optional<ButtonType> result = alert.showAndWait();
+
+        String message;
+        if(result.get().getButtonData() == ButtonBar.ButtonData.YES){
+            try {
+                ProductRepository.deleteProduct(product.getId().toString());
+                tableList.remove(product);
+                message = "Deleted";
+            }
+            catch (Exception e) {
+                message = "Cannot delete";
+            }
+        } else {
+            message = "Cancel delete";
+        }
+        raiseAlert(message);
+    }
+
+    private void raiseAlert(String alertText){
+        Alert alert = new Alert(Alert.AlertType.INFORMATION);
+        alert.setTitle("Information");
+        alert.setHeaderText("Notification");
+        alert.setContentText(alertText);
+        alert.show();
     }
     private void setTableView() throws IOException {
         products = ProductRepository.getProductsApi();
         tableList = FXCollections.observableArrayList(products);
         tableView.setItems(tableList);
     }
-    private void setColumn() throws IOException {
+    private void setColumn(){
         id.setCellValueFactory(new PropertyValueFactory<>("id"));
         name.setCellValueFactory(new PropertyValueFactory<>("name"));
         price.setCellValueFactory(new PropertyValueFactory<>("price"));
